website/requirements.txt
Flask==1.1.2
keras==2.11.0
matplotlib==3.5.2
numpy==1.21.5
pandas==1.4.4
pandas_datareader==0.10.0
python_dateutil==2.8.2
scikit_learn==1.2.1
tensorflow==2.11.0
tensorflow_intel==2.11.0
tqdm==4.64.1
yfinance==0.2.11
